extract notification service to a abstract service
