+++ src/server/plugins/engine/services/index.js
@@ -1,2 +1,3 @@
 export * as formsService from '~/src/server/plugins/engine/services/formsService.js'
 export * as formSubmissionService from '~/src/server/plugins/engine/services/formSubmissionService.js'
+export * as outputService from '~/src/server/plugins/engine/services/notifyService.js'
