+++ src/server/types.ts
@@ -5,7 +5,13 @@ import {
   type SubmitResponsePayload
 } from '@defra/forms-model'
 
-import { type FormStatus } from '~/src/server/routes/types.js'
+import { type FormModel } from './plugins/engine/models/index.js'
+import { type DetailItem } from './plugins/engine/models/types.js'
+
+import {
+  type FormRequestPayload,
+  type FormStatus
+} from '~/src/server/routes/types.js'
 
 export interface FormsService {
   getFormMetadata: (slug: string) => Promise<FormMetadata>
@@ -26,6 +32,7 @@ export interface FormSubmissionService {
 export interface Services {
   formsService: FormsService
   formSubmissionService: FormSubmissionService
+  outputService: OutputService
 }
 
 export interface RouteConfig {
@@ -33,4 +40,14 @@ export interface RouteConfig {
   formFilePath?: string
   enforceCsrf?: boolean
   services?: Services
+}
+
+export interface OutputService {
+  submit: (
+    request: FormRequestPayload,
+    model: FormModel,
+    emailAddress: string,
+    items: DetailItem[],
+    submitResponse: SubmitResponsePayload
+  ) => Promise<void>
 }
